+++ datajunction-server/datajunction_server/api/access/authentication/basic.py
@@ -11,11 +11,11 @@ from sqlmodel import Session, select
 
 from datajunction_server.constants import AUTH_COOKIE, LOGGED_IN_FLAG_COOKIE
 from datajunction_server.errors import DJError, DJException, ErrorCode
-from datajunction_server.internal.authentication.basic import (
+from datajunction_server.internal.access.authentication.basic import (
     get_password_hash,
     validate_user_password,
 )
-from datajunction_server.internal.authentication.tokens import create_token
+from datajunction_server.internal.access.authentication.tokens import create_token
 from datajunction_server.models.user import OAuthProvider, User
 from datajunction_server.utils import get_session
