+++ datajunction-server/datajunction_server/api/access/authentication/google.py
@@ -18,9 +18,12 @@ from starlette.responses import RedirectResponse
 
 from datajunction_server.constants import AUTH_COOKIE, LOGGED_IN_FLAG_COOKIE
 from datajunction_server.errors import DJException
-from datajunction_server.internal.authentication.basic import get_password_hash
-from datajunction_server.internal.authentication.google import flow, get_authorize_url
-from datajunction_server.internal.authentication.tokens import create_token
+from datajunction_server.internal.access.authentication.basic import get_password_hash
+from datajunction_server.internal.access.authentication.google import (
+    flow,
+    get_authorize_url,
+)
+from datajunction_server.internal.access.authentication.tokens import create_token
 from datajunction_server.models.user import OAuthProvider, User
 from datajunction_server.utils import Settings, get_session, get_settings
