+++ datajunction-server/datajunction_server/api/catalogs.py
@@ -12,7 +12,7 @@ from sqlmodel import Session, select
 from datajunction_server.api.engines import EngineInfo, get_engine
 from datajunction_server.api.helpers import get_catalog_by_name
 from datajunction_server.errors import DJException
-from datajunction_server.internal.authentication.http import SecureAPIRouter
+from datajunction_server.internal.access.authentication.http import SecureAPIRouter
 from datajunction_server.models.catalog import Catalog, CatalogInfo
 from datajunction_server.utils import get_session, get_settings
