+++ datajunction-server/datajunction_server/api/cubes.py
@@ -7,7 +7,7 @@ from fastapi import Depends
 from sqlmodel import Session
 
 from datajunction_server.api.helpers import get_node_by_name
-from datajunction_server.internal.authentication.http import SecureAPIRouter
+from datajunction_server.internal.access.authentication.http import SecureAPIRouter
 from datajunction_server.models.cube import CubeRevisionMetadata
 from datajunction_server.models.node import NodeType
 from datajunction_server.utils import get_session, get_settings
