+++ datajunction-server/datajunction_server/api/metrics.py
@@ -13,9 +13,9 @@ from sqlmodel import Session, select
 from datajunction_server.api.helpers import get_node_by_name
 from datajunction_server.api.nodes import list_nodes
 from datajunction_server.errors import DJError, DJException, ErrorCode
-from datajunction_server.internal.authentication.http import SecureAPIRouter
+from datajunction_server.internal.access.authentication.http import SecureAPIRouter
+from datajunction_server.internal.access.authorization import validate_access
 from datajunction_server.models import User, access
-from datajunction_server.models.access import validate_access
 from datajunction_server.models.metric import Metric
 from datajunction_server.models.node import DimensionAttributeOutput, Node, NodeType
 from datajunction_server.sql.dag import get_shared_dimensions
