+++ datajunction-server/datajunction_server/internal/access/authentication/google.py
@@ -13,7 +13,7 @@ from google.auth.external_account_authorized_user import Credentials
 from sqlmodel import select
 
 from datajunction_server.errors import DJException
-from datajunction_server.internal.authentication.basic import get_password_hash
+from datajunction_server.internal.access.authentication.basic import get_password_hash
 from datajunction_server.models.user import OAuthProvider, User
 from datajunction_server.utils import get_session, get_settings
